Added a checker that returns a solution score.

+++ ride.hpp
@@ -6,7 +6,7 @@
 struct Ride
 {
 	Point from, to;
-	unsigned int start, end;	
+	unsigned long long start, end;	
 	long long dist;
 
 	Ride( const Point &, const Point &,
